# archive/Workshed_Loom_Starter_Pack/openapi.yaml
openapi: 3.1.0
info:
  title: RABS-POC Workshed + Loom API
  version: 0.1.0
servers:
  - url: http://localhost:3009/api
paths:
...
  /loom/instances:
    get:
      summary: Read loom instances by date range
      parameters:
        - in: query
          name: startDate
          required: true
          schema: { type: string, format: date }
        - in: query
          name: endDate
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Instances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LoomInstance' }
...
components:
  schemas:
...
    LoomInstance:
      type: object
      properties:
        id: { type: string, format: uuid }
        template_id: { type: string, format: uuid }
        date: { type: string, format: date }
        status: { type: string }
        instance_staff_required: { type: integer }
        instance_vehicles_required: { type: integer }
...